S129148 upped version to 0.9.8, refactored processing of changesets to accomodate Pipeline builds that use changeSets property, moved this processing into a extractChangeSetInformation method, commented out SVN related tests due to inconsistency of json between diff Subversion plugin versions.
