+++ bldeif/bld_connector_runner.py
@@ -26,7 +26,7 @@ from bldeif.utils.eif_exception import OperationalError, logAllExceptions
 ############################################################################################################
 
 ARCHITECTURE_ACRONYM = 'eif'
-__version__ = "0.9.7"
+__version__ = "0.9.8"
 
 EXISTENCE_PROCLAMATION = """
 ************************************************************************************************************************
