+++ bldeif/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.9.7'
+__version__ = '0.9.8'
