+++ config/pipe.yml
@@ -1,4 +1,3 @@
-
 ---
 JenkinsBuildConnector:
 
@@ -14,16 +13,13 @@ JenkinsBuildConnector:
         Port      :  8080
         API_Token :  e008e30c73820b7eeb097ae1f1fa1dd8
         Username  :  jenkins
-        Password  :  None
-        MaxDepth  :  5
+        Password  :  encoded-T-m-9-u-Z-Q-=-=
+        MaxDepth  :  4
         AgileCentral_DefaultBuildProject :  Jenkins
 
         Jobs:
-            - Job: torment
-            - Job: bluestem
-            - Job: pillage-and-plunder
-            - Job: black-swan
-            - Job: Parkour
+            - Job: pipe dream
+            - Job: Dear Madame
 
     Service:
         Preview       : False
