+++ config/sample.yml
@@ -17,6 +17,7 @@ JenkinsBuildConnector:
         Password : somepassword
         API_Token: 320ca9ae9408d099183aa052ff3199c2
         # to get an API_Token, nav browser to  http://server:port/user/<username>/configure
+        MaxDepth  :  5  # specifies how many folder levels will be supported
         AgileCentral_DefaultBuildProject: Your Project 0
         
         Views:
