+++ config/trumpkin.yaml
@@ -13,7 +13,7 @@ JenkinsBuildConnector:
         Port      :  8080
         API_Token :  e008e30c73820b7eeb097ae1f1fa1dd8
         Username  :  jenkins
-        Password  :  encoded-T-m-9-u-Z-Q-=-=
+        Password  :  None
         MaxDepth  :  5
         AgileCentral_DefaultBuildProject :  Jenkins
